[cis] Refine code
